rafs: move RAFSv6 builder related code into a dedicated file
Move RAFSv6 builder related code into a dedicated file.
