+++ rafs/src/builder/core/mod.rs
@@ -12,3 +12,5 @@ pub(crate) mod node;
 pub(crate) mod overlay;
 pub(crate) mod prefetch;
 pub(crate) mod tree;
+pub(crate) mod v5;
+pub(crate) mod v6;
